Moved BasicImageParts to its own file
